filelimit to int8

+++ settings.go
@@ -32,7 +32,7 @@ const (
 type Settings struct {
 	file      string
 	Files     `yaml:"files,omitempty" json:"files,omitempty"`
-	FileLimit int64  `yaml:"flimit,omitempty" json:"flimit,omitempty"`
+	FileLimit int8  `yaml:"flimit,omitempty" json:"flimit,omitempty"`
 	Legacy    Legacy `yaml:"legacy" json:"legacy"`
 	Recovery  bool   `yaml:"recovery,omitempty" json:"recovery,omitempty"`
 }
